scaffolding

+++ geometry.c
@@ -14,22 +14,22 @@ struct Vec2d subtract(struct Vec2d v1, struct Vec2d v2) {
     return vs;
 }
 
-struct Vec2d multByScalar(struct Vec2d v, float scalar) {
+struct Vec2d multByScalar(struct Vec2d v, int scalar) {
     v.x *= scalar;
     v.y *= scalar;
     return v;
 }
 
-struct Vec2d divideByScalar(struct Vec2d v, float scalar) {
+struct Vec2d divideByScalar(struct Vec2d v, int scalar) {
     v.x /= scalar;
     v.y /= scalar;
     return v;
 }
 
 struct Vec2d normalize(struct Vec2d v) {
-    float l = Vec2dLen(v);
+    double l = Vec2dLen(v);
     if (l > 1) {
-        float inverse = 1 / l;
+        int inverse = 1 / l;
         struct Vec2d norm;
         norm = multByScalar(v, inverse);
         return norm;
